refactor: move data fetching from main to cardlist component
