+++ src/components/Details/style.css
@@ -1,0 +1,19 @@
+.overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 1;
+}
+
+.details {
+  position: relative;
+  z-index: 2;
+  width: 100%;
+  height: 100%;
+  padding: 20px 30px;
+  background-color: #464646;
+  border-radius: 10px;
+  overflow: hidden;
+}
