+++ src/index.css
@@ -20,7 +20,7 @@ body {
 }
 
 .main,
-.details,
+.aside,
 .container-404 {
   height: calc(100vh - 50px);
 }
@@ -44,23 +44,8 @@ body {
   text-align: center;
 }
 
-.details {
-  position: relative;
-  z-index: 2;
+.aside {
   width: 100%;
-  padding: 20px 30px;
-  background-color: #464646;
-  border-radius: 10px;
-  overflow: hidden;
-}
-
-.overlay {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 1;
 }
 
 .container-404 {
